dashboard/requirements.txt
pyserial==3.4
matplotlib==3.3.2
influxdb==5.3.1